Use dec10216 from msg_base.py both with HRIT- and NATIVE-readers.
